🔨 Refactor: nav-bar 폴더 위치 수정
