+++ src/components/common/index.ts
@@ -1,3 +1,4 @@
+export { default as NavBar } from "../nav-bar";
 export { default as AddListModalButton } from "./buttons/add-list-modal-button";
 export { default as Button } from "./buttons/button";
 export { default as CheckButton } from "./buttons/check-button";
@@ -14,6 +15,5 @@ export { default as motion } from "./framer-motion/motion";
 export { default as IconButton } from "./icon-button";
 export { default as Modal } from "./modal";
 export { default as OneInputForm } from "./modal/one-input-form";
-export { default as NavBar } from "./nav-bar";
 export { default as SidePage } from "./side-page";
 export { default as Toast } from "./toast";
