+++ src/components/nav-bar/user-dropdown.tsx
@@ -6,7 +6,7 @@ import { useToggle } from "@/hooks";
 import { IconUser } from "@/public/assets/icons";
 import { User } from "@/types/user";
 
-import DropDown from "../drop-down";
+import DropDown from "../common/drop-down";
 
 interface UserDropdownProps {
   user: User;
